Refactored the all functionalities of OnlineScreen into OnlineLibraryViewModel, and removed the dependency on ZimManageViewModel.
* Refactored the UI to use OnlineLibraryViewModel instead of ZimManageViewModel.
* Refactored AppProgressListenerProvider according to this change.
