+++ app/src/main/java/org/kiwix/kiwixmobile/nav/destination/library/online/OnlineCategoryDialogScreen.kt
@@ -50,7 +50,7 @@ import org.kiwix.kiwixmobile.core.ui.theme.KiwixDialogTheme
 import org.kiwix.kiwixmobile.core.utils.ComposeDimens
 import org.kiwix.kiwixmobile.language.LoadingScreen
 import org.kiwix.kiwixmobile.language.ShowErrorMessage
-import org.kiwix.kiwixmobile.nav.destination.library.online.viewmodel.Action.Select
+import org.kiwix.kiwixmobile.nav.destination.library.online.viewmodel.CategoryViewModel.Action.Select
 import org.kiwix.kiwixmobile.nav.destination.library.online.viewmodel.CategoryListItem
 import org.kiwix.kiwixmobile.nav.destination.library.online.viewmodel.CategoryListItem.CategoryItem
 import org.kiwix.kiwixmobile.nav.destination.library.online.viewmodel.CategoryListItem.HeaderItem
